% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIGDM_prob.r
\name{ZIGDM_prob}
\alias{ZIGDM_prob}
\title{Calculate probability based on zero inflated generalized Dirichlet-multinomial distribution}
\usage{
ZIGDM_prob(Y, Xc, Xa, Xb, c0, alpha0, beta0, tol = 1e-04, max.iter = 1000)
}
\arguments{
\item{Y}{Count table of microbiome}

\item{Xc}{design matrices for probability of absence model. The first column should contain "1" for the intercept. For intercept only model, the three matrices are simply contain vectors of 1.}

\item{Xa}{design matrices for mean model.  The first column should contain "1" for the intercept. For intercept only model, the three matrices are simply contain vectors of 1.}

\item{Xb}{design matrices for dispersion model. The first column should contain "1" for the intercept. For intercept only model, the three matrices are simply contain vectors of 1.}

\item{c0}{c0: initial values for regression coefficients "c" organized in a matrix (dim: K x d, K+1 is the #taxa (i.e. K+1 = ncol(Y)), d is the number of columns in the corresponding design matrix)}

\item{alpha0:}{initial values for regression coefficients "alpha" organized in a matrix (dim: K x d)}

\item{beta0:}{initial values for regression coefficients "beta" organized in a matrix (dim: K x d)}

\item{tol:}{convergence tolerance}

\item{max.iter:}{maximum number of iterations}
}
\value{
A vector containning the probablity of multinomial distribution
}
\description{
This function estimates parameters from zero inflated generalized Dirichlet-multinomial distribution, and test the pro.
}
